static utility renames
