+++ SimPEG/electromagnetics/static/induced_polarization/__init__.py
@@ -1,4 +1,4 @@
 from .simulation import Problem3D_CC, Problem3D_N
 from .simulation_2d import Problem2D_CC, Problem2D_N
 from .survey import Survey, from_dc_to_ip_survey
-from .Run import run_inversion
+from .run import run_inversion
