+++ SimPEG/electromagnetics/static/resistivity/__init__.py
@@ -11,4 +11,4 @@ from .fields_2d import Fields_ky, Fields_ky_CC, Fields_ky_N
 from .boundary_utils import getxBCyBC_CC
 from . import utils
 from .IODC import IO
-from .Run import run_inversion
+from .run import run_inversion
