+++ SimPEG/electromagnetics/static/resistivity/run.py
