+++ SimPEG/electromagnetics/static/spectral_induced_polarization/run.py
