+++ SimPEG/electromagnetics/static/utils/__init__.py
@@ -1,1 +1,1 @@
-from .StaticUtils import *
+from .static_utils import *
