+++ SimPEG/electromagnetics/static/utils/static_utils.py
