+++ tutorials/temporary/plot_3a_dcr_sounding_layers_inv.py
@@ -34,7 +34,7 @@ from SimPEG import (maps, data, data_misfit, regularization,
     optimization, inverse_problem, inversion, directives
     )
 from SimPEG.electromagnetics.static import resistivity as dc
-from SimPEG.electromagnetics.static.utils.StaticUtils import plot_layer
+from SimPEG.electromagnetics.static.utils.static_utils import plot_layer
 
 # sphinx_gallery_thumbnail_number = 2
